requirements.txt
aiogram==3.5.0
SQLAlchemy==2.0.27
pydantic==2.5.3
python-dotenv==1.0.0
geopy